feat : topRecord controller 통합테스트 완료
