+++ src/main/kotlin/chaegang/pl_api/domain/topRecord/TopRecord.kt
@@ -2,29 +2,11 @@ package chaegang.pl_api.domain.topRecord
 
 import jakarta.persistence.*
 
-/*
-    val name: String?,
-    val total: Float?,
-    val squat: Float?,
-    val bench: Float?,
-    val deadlift: Float?,
-    val sex: String?,
-    val bodyWeight: Double?,
-    val age: Float?,
-    val dots: Double?,
-    val wilks: Double?,
-    val glossbrenner: Double?,
-    val goodlift: Double?,
-    val tested: Boolean?,
-    val sanctioned: Boolean?,
-    val date: LocalDate?,
-    val federationName: String?,
-    val parentFederationName: String?,
- */
 @Entity
 class TopRecord (
     @Id
     val name: String,
+    val equipment: String? = null,
     val total: Float? = null,
     val squat: Float? = null,
     val bench: Float? = null,
