+++ src/main/kotlin/chaegang/pl_api/domain/topRecord/dto/TopRecordQueryResult.kt
@@ -4,6 +4,7 @@ import java.time.LocalDate
 
 class TopRecordQueryResult (
     val name: String?,
+    val equipment: String?,
     val total: Float?,
     val squat: Float?,
     val bench: Float?,
